requirements.txt
absl-py==0.13.0
chex==0.0.8
dill==0.3.4
immutabledict==2.0.0
jax==0.2.20
jaxline==0.0.3
numpy~=1.19.2
optax==0.0.9
Pillow==8.3.1
scipy==1.7.0
six~=1.15.0
tensorflow-datasets==4.3.0
tensorflow-probability==0.13.0
tensorflow==2.5.0
tensorflow-gan==2.0.0

